+++ src/types.ts
@@ -4,9 +4,16 @@ export type WpSite = {
 };
 
 
+export type WpGraphQlSort = {
+  fieldName: string,
+  direction: string,
+}
+
+
 export type WpGraphQlRequest = {
   pageSize: number;
   categoryId?: number;
   status?: string;
   fields?: string[];
+  orderBy?: WpGraphQlSort
 }